+++ src/AssemblyInfo.cs
@@ -2,5 +2,5 @@ using System.Reflection;
 
 [assembly: AssemblyTitle("AFKManager")]
 [assembly: AssemblyProduct("AFKManager")]
-[assembly: AssemblyVersion("0.1.0.0")]
-[assembly: AssemblyFileVersion("0.1.0.0")]
+[assembly: AssemblyVersion("0.2.0.0")]
+[assembly: AssemblyFileVersion("0.2.0.0")]